requirements.txt
chainlit
SpeechRecognition
PyAudio
pydub
numpy
openai
google-genai
python-dotenv
langchain
langchain-tavily
langchain-core
langchain-community
langchain-ollama
langchain-chroma
langchain_openai
langchain-google-genai
PyPDF2
python-docx
pillow
scikit-learn
beautifulsoup4
html2text
duckduckgo-search
ddgs
librosa
scipy
noisereduce
asyncpg
asyncmy
SQLAlchemy
aiohttp
azure-identity
azure-storage-file-datalake
greenlet
weasyprint
markdown